requirements.txt
python-dotenv
slack_sdk
feedparser
